Ites implemented as switches.  Supported in logic. (#163)
* Ites supported in logic.

* Ites implemented as switches.

* Ites supported in logic.

* Efficient data structure for ite switches.

* Space-efficient ite dag traversal

* All Ites are handled similarly (no special handling for Boolean Ites in cnfization)

+++ src/pterms/Pterm.h
@@ -96,7 +96,7 @@ class Pterm {
     Pterm operator= (Pterm&) = delete;
     Pterm operator= (Pterm&&) = delete;
 
-    int      size        ()          const   { return header.size; }
+    int      size        ()          const   { return static_cast<int>(header.size); }
 
     const PTRef& operator [] (int i) const   { assert(i < size()); return args[i]; }
     PTRef&       operator [] (int i)         { assert(i < size()); return args[i]; }
